+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Health check endpoints are available at `/startupz`, `/readyz`,
   `/livez`, as well as `/health` and `/healthz`.
+- Prometheus/OpenMetrics metrics is available at `/metrics`.
 
 ### Internal